mobile: split TransfersScreen.tsx (1195L) into focused sibling files
Single-file screen had 22 inline sub-components spanning fetching,
list rendering, card layout, comparison expansion, and empty states.
At that size new readers can't navigate by jumping to definitions.

Split into mobile/src/screens/Transfers/:
- index.tsx (311L) — TransfersScreen + SuggestionsView + Body +
  ControlsRow. The fetch-and-state-branch orchestration layer.
- SuggestionsList.tsx (112L) — FlatList + Header + bundleKey.
  Owns the one-card-expanded-at-a-time state.
- BundleCard.tsx (135L) — BundleCard + BundleSummary + HitPill.
- MoveRow.tsx (141L) — PlayerBlock + CenterBadge + BankDeltaPill.
  BankDeltaPill is here rather than its own file because it's used
  by CenterBadge here AND BundleSummary in BundleCard — exporting
  it from the move-row module is the natural seam.
- CompareTable.tsx (103L) — CompareTable + Row + ToneCell + fmt.
- EmptyStates.tsx (57L) — NoTeamIdState + PicksNotFoundState +
  MessageState.
- styles.ts (374L) — shared makeStyles block, co-located rather
  than split per-component because cross-references between cards,
  pills, and the compare table are dense.

Navigation stack now imports from '../../screens/Transfers' (folder
index). No behavior changes: stateful interactions (LayoutAnimation,
expand/collapse, horizon switch, position filter) preserved.
